cross db compatibility

+++ models/stg_linkedin__campaign_history.sql
@@ -12,7 +12,7 @@ with base as (
         campaign_group_id,
         created_time as created_at,
         name as campaign_name,
-        cast(version_tag as INT64) as version_tag
+        cast(version_tag as numeric) as version_tag
     from base
 
 ), valid_dates as (
